sender_requirements.txt
croniter
gspread
python-dotenv
flask
pydantic
git+https://github.com/leshchenko1979/tg.git
tgcrypto
pyyaml
